requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
motor>=3.3
python-dotenv>=1.0
apify-client>=1.6